Add os_type for Juju Mixin (#764)
* Add JujuMixin class

* Add Juju-based Aws VM

* Juju support

Add support for running Juju on AWS as an os_type
Create a virtual machine for juju controller
Add new virtual machines to Juju’s control

* Add support for running the Cassandra benchmark with the Juju os_type

* Resolve merge conflicts

* Fix papercuts introduced by merge of master.

* pep8 fixes

* Remove pkb log

* Master sync

Sync the current work from master

* Don't add Juju as a cloud

* Git shenanigans

* Remove merge conflict artifact

* Fix merge errors

* Really fix the failing test this time.

* Remove juju-specific bits

* Move cassandra-stress binary location to the cassandra package

* Move all juju-specific code out of the benchmark and into this package

* Initial handling for units in an error state, to prevent blocking

* Cleanup; increase timeout between status checks from 15 -> 30s

* Fix flake8 error

* Document running benchmarks with juju

* Fix typos

* Remove trailing Whitespace

* Use the vmspec's os_type instead of relying on the OS flag

* Gandalfing unnecessary pass's

* Don't eat exceptions

* Pass the vm to cassandra.GetCassandraStressPath so it can be overridden as necessary

* Rename Juju-specific methods to follow CamelCase
Add timeout handling to JujuWait
Remove commented-out code

* Compare against vm.OS_TYPE instead of FLAGS.os_type
Update new Juju method names
Increase timeout
Add vm as a param to GetCassandraStressPath so the path can be changed based on OS_TYPE

* Cleanup os_type check

* Refactor ConstructVirtualMachines to support Juju across multiple clouds and groups

* Small refactor of JujuInstall

* Fix JujuMixin spacing to match the rest of the code
Add proper timeout handling to JujuWait

* Use os_types.JUJU constant

* Fall back to AptInstall if JujuInstall isn't available in the package

* Remove cruft

* Remove errant code

* Add Juju-specific exceptions

* JujuWait() will now raise an exception (using @vm_util.Retry) on error or timeout

* Remove enumerate(), since the use of idx ended up being not necessary

* Rename ConstructVirtualMachine to ConstructVirtualMachineGroup, to better reflect it's function

* Use os_types.JUJU in lieu of local constant
Short-circuit based on os_types.JUJU

* Nip dead code

* Use vm_util.NamedTemporaryFile

* Fix variable name

* fix lint

* Remove unused 'tempfile' import

* Track vm_group_specs so Juju can deploy the exact number of machines needed in the workers group

* Don't taint non-juju vm_groups with the juju controller.

* Update documentation of BaseVirtualMachine's ip_address attribute

* Track the vm_group each unit belongs to

* Specify the vm_group each service should be deployed to

* Add more documentation
Add support for placing a service in a specific vm_group

* Fix indentation

* Refactor JujuDeploy: used cached machine id from add-machine output

* Change argument order for JujuDeploy

* Fix parsing of machine_id

* Create a separate cassandra-stress tool package

* Refactor to separate cassandra and cassandra-stress packages

* Refactor JujuMixin.Install()

* Flake8 fix

* Flake8 fix

* Add method to get the path to cassandra-cli, since that changes if cassandra is installed by Juju

* Downgrade the Cassandra charm from 2.2.x to 2.1.x, so that the cassandra-cli command is available (adds compatibility for ycsb benchmark)

* Add helper method for constructing Juju controller

* Apply style guidelines. Remove extraneous checks for vm.isController.

* Apply style guide

* Remove legacy units parameter

* Don't camelcase variables

* Revert line-break (previous lint error?)

* styleguide!

* Remove whitespace from blank line

* Add TODO for uninstall/cleanup functionality

* Remove unused class variable vm_group_specs

* Fix double/single quotes to match style guide

* More pesky double quotes

+++ perfkitbenchmarker/linux_packages/cassandra.py
@@ -26,6 +26,8 @@ import time
 
 from perfkitbenchmarker import data
 from perfkitbenchmarker import errors
+from perfkitbenchmarker import flags
+from perfkitbenchmarker import os_types
 from perfkitbenchmarker import vm_util
 from perfkitbenchmarker.linux_packages.ant import ANT_HOME_DIR
 
@@ -48,6 +50,8 @@ CLUSTER_START_TRIES = 10
 CLUSTER_START_SLEEP = 60
 # Time, in seconds, to sleep between node starts.
 NODE_START_SLEEP = 5
+
+FLAGS = flags.FLAGS
 
 
 def CheckPrerequisites():
@@ -90,6 +94,40 @@ def AptInstall(vm):
   _Install(vm)
 
 
+def JujuInstall(vm, vm_group_name):
+  """Installs the Cassandra charm on the VM."""
+  vm.JujuDeploy('cs:trusty/cassandra', vm_group_name)
+
+  # The charm defaults to Cassandra 2.2.x, which has deprecated
+  # cassandra-cli. Specify the sources to downgrade to Cassandra 2.1.x
+  # to match the cassandra benchmark(s) expectations.
+  sources = ['deb http://www.apache.org/dist/cassandra/debian 21x main',
+             'ppa:openjdk-r/ppa',
+             'ppa:stub/cassandra']
+
+  keys = ['F758CE318D77295D',
+          'null',
+          'null']
+
+  vm.JujuSet('cassandra', [
+             # Allow authentication from all units
+             'authenticator=AllowAllAuthenticator',
+             'install_sources="[%s]"' %
+             ', '.join(map(lambda x: "'" + x + "'", sources)),
+             'install_keys="[%s]"'
+             % ', '.join(keys)
+             ])
+
+  # Wait for cassandra to be installed and configured
+  vm.JujuWait()
+
+  for unit in vm.units:
+    # Make sure the cassandra/conf dir is created, since we're skipping
+    # the manual installation to /tmp/pkb.
+    remote_path = posixpath.join(CASSANDRA_DIR, 'conf')
+    unit.RemoteCommand('mkdir -p %s' % remote_path)
+
+
 def Configure(vm, seed_vms):
   """Configure Cassandra on 'vm'.
 
@@ -117,6 +155,9 @@ def Start(vm):
   Args:
     vm: The target vm. Should already be configured via 'Configure'.
   """
+  if vm.OS_TYPE == os_types.JUJU:
+    return
+
   vm.RemoteCommand(
       'nohup {0}/bin/cassandra -p "{1}" 1> {2} 2> {3} &'.format(
           CASSANDRA_DIR, CASSANDRA_PID, CASSANDRA_OUT, CASSANDRA_ERR))
@@ -124,6 +165,9 @@ def Start(vm):
 
 def Stop(vm):
   """Stops Cassandra on 'vm'."""
+  if vm.OS_TYPE == os_types.JUJU:
+    return
+
   vm.RemoteCommand('kill $(cat {0})'.format(CASSANDRA_PID),
                    ignore_failure=True)
 
@@ -153,6 +197,9 @@ def CleanNode(vm):
   Args:
     vm: VirtualMachine. VM to clean.
   """
+  if vm.OS_TYPE == os_types.JUJU:
+    return
+
   data_path = posixpath.join(vm.GetScratchDir(), 'cassandra')
   vm.RemoteCommand('rm -rf {0}'.format(data_path))
 
@@ -162,6 +209,26 @@ def _StartCassandraIfNotRunning(vm):
   if not IsRunning(vm):
     logging.info('Retrying starting cassandra on %s', vm)
     Start(vm)
+
+
+def GetCassandraCliPath(vm):
+  if vm.OS_TYPE == os_types.JUJU:
+    # Replace the stock CASSANDRA_CLI so that it uses the binary
+    # installed by the cassandra-stress charm.
+    return '/usr/bin/cassandra-cli'
+
+  return posixpath.join(CASSANDRA_DIR, 'tools', 'bin',
+                        'cassandra-cli')
+
+
+def GetCassandraStressPath(vm):
+  if vm.OS_TYPE == os_types.JUJU:
+    # Replace the stock CASSANDRA_STRESS so that it uses the binary
+    # installed by the cassandra-stress charm.
+    return '/usr/bin/cassandra-stress'
+
+  return posixpath.join(CASSANDRA_DIR, 'tools', 'bin',
+                        'cassandra-stress')
 
 
 def GetNumberOfNodesUp(vm):
@@ -187,6 +254,11 @@ def StartCluster(seed_vm, vms):
     vms: list of VirtualMachines. VMs *other than* seed_vm which should be
       started.
   """
+
+  if seed_vm.OS_TYPE == os_types.JUJU:
+    # Juju automatically configures and starts the Cassandra cluster.
+    return
+
   vm_count = len(vms) + 1
 
   # Cassandra setup
